Add customization component for about page texts
Refs #327.
